pyproject.toml
[project]
name = "mlflow-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mlflow>=3.5.1",
    "numpy>=2.3.4",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.2.250926",
    "ruff>=0.14.3",
]

[tool.ruff]
line-length = 70 # 코드 한 줄의 최대 길이
ignore = ["F401"]

[tool.ruff.lint]
select = ["E", "F", "W"] # 모든 경고 선택
